moved uploading to db from stats to bedboss
