+++ bedboss/bedstat/bedstat.py
@@ -3,17 +3,12 @@ import json
 import os
 import requests
 import pypiper
-import bbconf
 import logging
 from geniml.io import RegionSet
-from pephubclient import PEPHubClient
-from pephubclient.helpers import is_registry_path
-from ubiquerg import parse_registry_path
+
 
 from bedboss.const import (
     OUTPUT_FOLDER_NAME,
-    BED_FOLDER_NAME,
-    BIGBED_FOLDER_NAME,
     BEDSTAT_OUTPUT,
     OS_HG19,
     OS_HG38,
@@ -24,7 +19,6 @@ from bedboss.const import (
 )
 from bedboss.utils import download_file, convert_unit
 from bedboss.exceptions import OpenSignalMatrixException
-from bedboss.models import BedMetadata
 
 
 _LOGGER = logging.getLogger("bedboss")
@@ -32,81 +26,6 @@ _LOGGER = logging.getLogger("bedboss")
 SCHEMA_PATH_BEDSTAT = os.path.join(
     os.path.dirname(os.path.realpath(__file__)), "pep_schema.yaml"
 )
-
-BED_PEP_REGISTRY = "databio/allbeds:bedbase"
-
-
-def load_to_pephub(
-    pep_registry_path: str, bed_digest: str, genome: str, metadata: dict
-) -> None:
-    """
-    Load bedfile and metadata to PEPHUB
-
-    :param str pep_registry_path: registry path to pep on pephub
-    :param str bed_digest: unique bedfile identifier
-    :param str genome: genome associated with bedfile
-    :param dict metadata: Any other metadata that has been collected
-
-    :return None
-    """
-
-    if is_registry_path(pep_registry_path):
-        parsed_pep_dict = parse_registry_path(pep_registry_path)
-
-        # Combine data into a dict for sending to pephub
-        sample_data = {}
-        sample_data.update({"sample_name": bed_digest, "genome": genome})
-
-        metadata = BedMetadata(**metadata).model_dump()
-
-        for key, value in metadata.items():
-            # TODO: Confirm this key is in the schema
-            # Then update sample_data
-            sample_data.update({key: value})
-
-        try:
-            PEPHubClient().sample.create(
-                namespace=parsed_pep_dict["namespace"],
-                name=parsed_pep_dict["item"],
-                tag=parsed_pep_dict["tag"],
-                sample_name=bed_digest,
-                overwrite=True,
-                sample_dict=sample_data,
-            )
-
-        except Exception as e:  # Need more specific exception
-            _LOGGER.error(f"Failed to upload BEDFILE to PEPhub: See {e}")
-    else:
-        _LOGGER.error(f"{pep_registry_path} is not a valid registry path")
-
-
-def load_to_s3(
-    output_folder: str,
-    pm: pypiper.PipelineManager,
-    bed_file: str,
-    digest: str,
-    bigbed_file: str = None,
-) -> None:
-    """
-    Load bedfiles and statistics to s3
-
-    :param output_folder: base output folder
-    :param pm: pipelineManager object
-    :param bed_file: bedfile name
-    :param digest: bedfile digest
-    :param bigbed_file: bigbed file name
-    :return: NoReturn
-    """
-    command = f"aws s3 cp {os.path.join(output_folder, bed_file)} s3://bedbase/{BED_FOLDER_NAME}"
-    _LOGGER.info("Uploading to s3 bed file")
-    pm.run(cmd=command, lock_name="s3_sync_bed")
-    if bigbed_file:
-        command = f"aws s3 cp {os.path.join(output_folder, bigbed_file)} s3://bedbase/{BIGBED_FOLDER_NAME}"
-        _LOGGER.info("Uploading to s3 bigbed file")
-        pm.run(cmd=command, lock_name="s3_sync_bigbed")
-    command = f"aws s3 sync {os.path.join(output_folder, OUTPUT_FOLDER_NAME,BEDSTAT_OUTPUT, digest)} s3://bedbase/{OUTPUT_FOLDER_NAME}/{BEDSTAT_OUTPUT}/{digest} --size-only"
-    _LOGGER.info("Uploading to s3 bed statistic files")
-    pm.run(cmd=command, lock_name="s3_sync_bedstat")
 
 
 def get_osm_path(genome: str, out_path: str = None) -> Union[str, None]:
@@ -147,22 +66,16 @@ def get_osm_path(genome: str, out_path: str = None) -> Union[str, None]:
 
 def bedstat(
     bedfile: str,
-    bedbase_config: Union[str, bbconf.BedBaseConf],
     genome: str,
     outfolder: str,
+    bed_digest: str = None,
+    bigbed: str = None,
     ensdb: str = None,
     open_signal_matrix: str = None,
-    bigbed: str = None,
-    other_metadata: dict = None,
     just_db_commit: bool = False,
-    no_db_commit: bool = False,
-    force_overwrite: bool = False,
-    skip_qdrant: bool = True,
-    upload_s3: bool = False,
-    upload_pephub: bool = False,
     pm: pypiper.PipelineManager = None,
-    **kwargs,
-) -> str:
+    # **kwargs,
+) -> dict:
     """
     Run bedstat pipeline - pipeline for obtaining statistics about bed files
         and inserting them into the database
@@ -171,24 +84,16 @@ def bedstat(
     :param str bigbed: the full path to the bigbed file. Defaults to None.
         (bigbed won't be created and some producing of some statistics will
         be skipped.)
-    :param str bedbase_config: The path to the bedbase configuration file, or bbconf object
+    :param str bed_digest: the digest of the bed file. Defaults to None.
     :param str open_signal_matrix: a full path to the openSignalMatrix
         required for the tissue specificity plots
     :param str outfolder: The folder for storing the pipeline results.
     :param str genome: genome assembly of the sample
     :param str ensdb: a full path to the ensdb gtf file required for genomes
         not in GDdata
-    :param dict other_metadata: a dictionary of other metadata to pass
-    :param bool just_db_commit: whether just to commit the JSON to the database
-    :param bool no_db_commit: whether the JSON commit to the database should be
-        skipped
-    :param skip_qdrant: whether to skip qdrant indexing [Default: True]
-    :param bool force_overwrite: whether to overwrite the existing record
-    :param upload_s3: whether to upload the bed file to s3
-    :param bool upload_pephub: whether to push bedfiles and metadata to pephub (default: False)
     :param pm: pypiper object
 
-    :return: bed_digest: the digest of the bed file
+    :return: dict with statistics and plots metadata
     """
     # TODO why are we no longer using bbconf to get the output path?
     # outfolder_stats = bbc.get_bedstat_output_path()
@@ -198,12 +103,6 @@ def bedstat(
         os.makedirs(outfolder_stats)
     except FileExistsError:
         pass
-
-    # if bbconf is a string, create a bbconf object
-    if isinstance(bedbase_config, str):
-        bbc = bbconf.BedBaseConf(config_path=bedbase_config, database_only=True)
-    else:
-        bbc = bedbase_config
 
     # find/download open signal matrix
     if not open_signal_matrix or not os.path.exists(open_signal_matrix):
@@ -221,7 +120,8 @@ def bedstat(
     else:
         stop_pipeline = False
 
-    bed_digest = RegionSet(bedfile).identifier
+    if not bed_digest:
+        bed_digest = RegionSet(bedfile).identifier
     bedfile_name = os.path.split(bedfile)[1]
 
     fileid = os.path.splitext(os.path.splitext(bedfile_name)[0])[0]
@@ -277,132 +177,79 @@ def bedstat(
 
         pm.run(cmd=command, target=json_file_path)
 
-    # commit to the database if no_db_commit is not set
-    if not no_db_commit:
-        data = {}
-        if os.path.exists(json_file_path):
-            with open(json_file_path, "r", encoding="utf-8") as f:
-                data = json.loads(f.read())
-        if os.path.exists(json_plots_file_path):
-            with open(json_plots_file_path, "r", encoding="utf-8") as f_plots:
-                plots = json.loads(f_plots.read())
-        else:
-            plots = []
+    data = {}
+    if os.path.exists(json_file_path):
+        with open(json_file_path, "r", encoding="utf-8") as f:
+            data = json.loads(f.read())
+    if os.path.exists(json_plots_file_path):
+        with open(json_plots_file_path, "r", encoding="utf-8") as f_plots:
+            plots = json.loads(f_plots.read())
+    else:
+        plots = []
 
-        if not other_metadata:
-            other_metadata = {}
+    # unlist the data, since the output of regionstat.R is a dict of lists of
+    # length 1 and force keys to lower to correspond with the
+    # postgres column identifiers
+    data = {k.lower(): v[0] if isinstance(v, list) else v for k, v in data.items()}
+    data.update(
+        {
+            "bedfile": {
+                "path": bed_relpath,
+                "size": convert_unit(os.path.getsize(bedfile)),
+                "title": "Path to the BED file",
+            }
+        }
+    )
 
-        # unlist the data, since the output of regionstat.R is a dict of lists of
-        # length 1 and force keys to lower to correspond with the
-        # postgres column identifiers
-        data = {k.lower(): v[0] if isinstance(v, list) else v for k, v in data.items()}
+    if os.path.exists(os.path.join(bigbed, fileid + ".bigBed")):
         data.update(
             {
-                "bedfile": {
-                    "path": bed_relpath,
-                    "size": convert_unit(os.path.getsize(bedfile)),
-                    "title": "Path to the BED file",
+                "bigbedfile": {
+                    "path": bigbed_relpath,
+                    "size": convert_unit(
+                        os.path.getsize(os.path.join(bigbed, fileid + ".bigBed"))
+                    ),
+                    "title": "Path to the big BED file",
                 }
             }
         )
 
-        if os.path.exists(os.path.join(bigbed, fileid + ".bigBed")):
-            data.update(
-                {
-                    "bigbedfile": {
-                        "path": bigbed_relpath,
-                        "size": convert_unit(
-                            os.path.getsize(os.path.join(bigbed, fileid + ".bigBed"))
-                        ),
-                        "title": "Path to the big BED file",
-                    }
-                }
-            )
+        if not os.path.islink(os.path.join(bigbed, fileid + ".bigBed")):
+            digest = requests.get(
+                f"http://refgenomes.databio.org/genomes/genome_digest/{genome}"
+            ).text.strip('""')
 
-            if not os.path.islink(os.path.join(bigbed, fileid + ".bigBed")):
-                digest = requests.get(
-                    f"http://refgenomes.databio.org/genomes/genome_digest/{genome}"
-                ).text.strip('""')
-
-                data.update(
-                    {
-                        "genome": {
-                            "alias": genome,
-                            "digest": digest,
-                        }
-                    }
-                )
-        else:
             data.update(
                 {
                     "genome": {
                         "alias": genome,
-                        "digest": "",
+                        "digest": digest,
                     }
                 }
             )
+    else:
+        data.update(
+            {
+                "genome": {
+                    "alias": genome,
+                    "digest": "",
+                }
+            }
+        )
 
-        for plot in plots:
-            plot_id = plot["name"]
-            del plot["name"]
-            data.update({plot_id: plot})
+    for plot in plots:
+        plot_id = plot["name"]
+        del plot["name"]
+        data.update({plot_id: plot})
 
-        # deleting md5sum, because it is record_identifier
+    # deleting md5sum, because it is record_identifier
+    if "md5sum" in data:
         del data["md5sum"]
 
-        # add added_to_qdrant to the data
-        data["added_to_qdrant"] = False
-
-        # add other to dict in bb database (now we are using pephub for this purpose)
-        # data["other"] = other_metadata
-
-        bbc.bed.report(
-            record_identifier=bed_digest,
-            values=data,
-            force_overwrite=force_overwrite,
-        )
-
-    if upload_s3:
-        _LOGGER.info(f"Uploading '{bed_digest}' data to S3 ...")
-        load_to_s3(
-            os.path.abspath(outfolder), pm, bed_relpath, bed_digest, bigbed_relpath
-        )
-    else:
-        _LOGGER.info(
-            f"Skipping uploading '{bed_digest}' data to S3. 'upload_s3' is set to False. "
-        )
-
-    if not skip_qdrant:
-        _LOGGER.info(f"Adding '{bed_digest}' vector to Qdrant ...")
-
-        bbc.add_bed_to_qdrant(
-            bed_id=bed_digest,
-            bed_file=bedfile,
-            payload={"fileid": fileid},
-        )
-        bbc.bed.report(
-            record_identifier=bed_digest,
-            values={"added_to_qdrant": True},
-            force_overwrite=True,
-        )
-    else:
-        _LOGGER.info(
-            f"Skipping adding '{bed_digest}' vector to Qdrant, 'skip_qdrant' is set to True. "
-        )
-
-    if upload_pephub:
-        _LOGGER.info(f"Uploading metadata of '{bed_digest}' TO PEPhub ...")
-        load_to_pephub(
-            pep_registry_path=BED_PEP_REGISTRY,
-            bed_digest=bed_digest,
-            genome=genome,
-            metadata=other_metadata,
-        )
-    else:
-        _LOGGER.info(
-            f"Metadata of '{bed_digest}' is NOT uploaded to PEPhub. 'upload_pephub' is set to False. "
-        )
+    # add added_to_qdrant to the data
+    data["added_to_qdrant"] = False
 
     if stop_pipeline:
         pm.stop_pipeline()
-    return bed_digest
+
+    return data
