--- bedboss/bedstat/models.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel, ConfigDict, Field
-
-
-class BEDSTAT_RETURN(BaseModel):
-    """
-    Model of single namespace search result
-    """
-
-    ...
