--- bedboss/bedstat/pep_schema.yaml
@@ -1,79 +1,0 @@
-description: bedstat PEP schema
-
-properties:
-  samples:
-    type: array
-    items:
-      type: object
-      properties:
-        sample_name: 
-          type: string
-          db_commit: TRUE
-          description: "name of the sample, which is the name of the output BED file"
-        input_file_path:
-          type: string
-          db_commit: FALSE
-          description: "absolute path the file to convert"
-        output_file_path:
-          type: string
-          db_commit: FALSE
-          description: "absolute path the file to the output BED file (derived attribute)"
-        bigbed:
-          type: string
-          db_commit: FALSE
-          description: "dir path where the bigbed file stored (derived attribute)"
-        genome:
-          type: string
-          db_commit: TRUE
-          description: "organism genome code"
-        narrowpeak:
-          type: boolean
-          db_commit: TRUE
-          description: "binary number indicating whether the regions are narrow (transcription factor implies narrow, histone mark implies broad peaks)"
-        format:
-          type: string
-          db_commit: TRUE
-          description: "file format"
-          enum: ["bigWig", "bigBed", "bed", "wig", "bedGraph"]
-        cell_type:
-          type: string
-          db_commit: TRUE
-          description: "cell type code"
-        antibody:
-          type: string
-          db_commit: TRUE
-          description: "antibody used if ChIP-seq experiment"
-        description:
-          type: string
-          db_commit: TRUE
-          description: "freeform description of the sample"
-        exp_protocol:
-          type: string
-          db_commit: TRUE
-          description: "type of the experiment the file was generated in"
-        data_source:
-          type: string
-          db_commit: TRUE
-          description: "source of the sample, preferably a GSE* code"
-        treatment:
-          type: string
-          db_commit: TRUE
-          description: "freeform description of the sample treatment"
-        ensdb:
-          type: string
-          db_commit: FALSE
-          description: "path of gtf annotation for genomes not in GDdata"
-        fasta:
-          type: string
-          db_commit: FALSE
-          description: "path of for genomes not in GDdata"
-        open_signal_matrix:
-          type: string
-          db_commit: FALSE
-          description: "path of for the open signal matrixm file for the given genome"
-      required:
-        - output_file_path
-        - genome
-        - sample_name
-required:
-  - samples
